[WEJBHTTP-141] Refactoring - Renaming JndiConstants to Constants.
